Move flake8-debugger violation
